+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ApplicationInsights Winstron Transport
 
+## 0.2.1
+
+- removed custom winston-transport typings in preperation for winston@3 release
+- fix for wrong `useRetryDiskCach` default when client options a present but the key is undefined
+
 ## 0.2.0
 - expose maxBatchIntervalMs and useRetryDiskCache options